Move configuration out of main.rs
All configuration utilities are moved into a separated module: config.rs
