+++ src/constants.rs
@@ -114,6 +114,7 @@ pub const LOG_EMITTER_TASK_COMMAND: &str = "COMMAND_TASK";
 pub const LOG_EMITTER_TASK_LUA: &str = "LUA_TASK";
 pub const LOG_EMITTER_EVENT_DBUS: &str = "DBUS_EVENT";
 pub const LOG_EMITTER_EVENT_FSCHANGE: &str = "FSCHANGE_EVENT";
+pub const LOG_EMITTER_EVENT_MANUAL: &str = "CMD_EVENT";
 pub const LOG_EMITTER_CONDITION_INTERVAL: &str = "INTERVAL_CONDITION";
 pub const LOG_EMITTER_CONDITION_BUCKET: &str = "BUCKET_CONDITION";
 pub const LOG_EMITTER_CONDITION_COMMAND: &str = "COMMAND_CONDITION";
