+++ src/event/manual_event.rs
@@ -62,7 +62,7 @@ impl ManualCommandEvent {
     pub fn new(name: &str) -> Self {
         log(
             LogType::Debug,
-            LOG_EMITTER_EVENT_FSCHANGE,
+            LOG_EMITTER_EVENT_MANUAL,
             LOG_ACTION_NEW,
             Some((name, 0)),
             LOG_WHEN_INIT,
